Moved database objects into domain package and removed old database objects
