+++ app/src/main/java/com/example/xjapan/karaoke2/Karaoke2.java
@@ -2,7 +2,7 @@ package com.example.xjapan.karaoke2;
 
 import android.app.Application;
 
-import com.example.xjapan.karaoke2.infra.db.dao.DAOInjector;
+import com.example.xjapan.karaoke2.domain.dao.DAOInjector;
 
 /**
  * Created by jmatsu on 2016/02/27.
