+++ app/src/main/java/com/example/xjapan/karaoke2/domain/dao/UserDAO.java
@@ -1,7 +1,7 @@
-package com.example.xjapan.karaoke2.infra.db.dao;
+package com.example.xjapan.karaoke2.domain.dao;
 
-import com.example.xjapan.karaoke2.infra.db.entity.OrmaDatabase;
-import com.example.xjapan.karaoke2.infra.db.entity.User;
+import com.example.xjapan.karaoke2.domain.entity.OrmaDatabase;
+import com.example.xjapan.karaoke2.domain.entity.User;
 
 /**
  * Created by jmatsu on 2016/02/27.
