+++ app/src/main/java/com/example/xjapan/karaoke2/presentation/activity/initialization/InitialUseActivity.java
@@ -17,25 +17,19 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
-import android.text.SpannableStringBuilder;
-import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 
-import com.example.xjapan.karaoke2.infra.db.entity.User;
+import com.example.xjapan.karaoke2.domain.entity.User;
 import com.example.xjapan.karaoke2.presentation.activity.MainActivity;
 import com.example.xjapan.karaoke2.R;
-import com.example.xjapan.karaoke2.infra.db.dao.UserDB;
-import com.example.xjapan.karaoke2.infra.api.AppClient;
 import com.example.xjapan.karaoke2.usecase.common.FailureCallback;
 import com.example.xjapan.karaoke2.usecase.common.RetrofitSuccessEvent;
 import com.example.xjapan.karaoke2.usecase.common.SuccessCallback;
 import com.example.xjapan.karaoke2.usecase.initialization.CreateUserUseCase;
 
-import retrofit.Callback;
 import retrofit.RetrofitError;
-import retrofit.client.Response;
 
 public class InitialUseActivity extends AppCompatActivity {
